created code base with empty functions on curves.
Created main function "search".
Working fine on assignment 1 search.

+++ include/lsh_discrete.h
@@ -13,7 +13,7 @@ private:
 
 public:
 
-    DiscreteLSHcurve(unsigned int N, unsigned int dimensions, float delta, unsigned int L=5);
+    DiscreteLSHcurve(unsigned int N, unsigned int dimensions, double delta, unsigned int L=5);
     ~DiscreteLSHcurve();
 };
